-Se modificaron los parametros y se agregaron nuevas validaciones del constructor del formulario frmSearchGuest. -Se documentaron parametro del ValidateHelper. -Se agregaron parametros, se documentaron y agregaron nuevos BR. -Se agregaron styles en el App.xaml. -Se agregaron parametros, se documentaron y agregaron metodos en las classes de Host. -Se estandarizo names de frmSearchGuest. -Se modifico algunos detalles de los formularios de Host.png -Se renombro el EnumSalesRoomType a EnumEntities. -Se volvio asincrono metodos del SisturHelper.

+++ IM.Base/Helpers/GridHelper.cs
@@ -373,19 +373,19 @@ namespace IM.Base.Helpers
     public static void ValidateEditNumber(ref int pNumber, bool pCancel, string pTitle, int pUpperBound, int pLowerBound, int pDefaultValue = 0, bool pValidateBounds = true)
     {
       // si se ingreso un valor
-      if (pNumber >= 0)
+      if ((int)pNumber >= 0)
       {
         // si se desea validar los limites
         if (pValidateBounds)
         {
           // validamos que no sea mayor al limite superior
-          if (pNumber > pUpperBound)
+          if ((int)pNumber > pUpperBound)
           {
             UIHelper.ShowMessage("Quantity can not be greater than " + pUpperBound, MessageBoxImage.Information, "Intelligence Marketing");
             pCancel = true;
           }
           // validamos que no sea menor al limite inferior
-          else if (pNumber < pLowerBound)
+          else if ((int)pNumber < pLowerBound)
           {
             UIHelper.ShowMessage("Quantity can not be lower than " + pLowerBound, MessageBoxImage.Information, "Intelligence Marketing");
             pCancel = true;
